+++ src/core.rs
@@ -16,5 +16,8 @@ pub use level::Level;
 mod proficiency_bonus;
 pub use proficiency_bonus::ProficiencyBonus;
 
+mod skill_proficiencies;
+pub use skill_proficiencies::{SkillLevel, SkillProficiencies};
+
 mod skill;
 pub use skill::Skill;