+++ src/core/proficiency_bonus.rs
@@ -2,7 +2,7 @@ use crate::core::Level;
 
 /// Represents a proficiency bonus.
 ///
-/// A proficiency bonus is a [`u8`] value in the range of `2..=9`.
+/// A proficiency bonus is a [`u8`] value in the range of `2..=18`.
 ///
 /// # Examples
 ///
@@ -14,6 +14,7 @@ use crate::core::Level;
 /// ```
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[repr(transparent)]
 pub struct ProficiencyBonus(u8);
 
 impl ProficiencyBonus {
@@ -21,18 +22,18 @@ impl ProficiencyBonus {
     pub const MIN: Self = Self(2);
 
     /// The maximum possible proficiency bonus.
-    pub const MAX: Self = Self(9);
+    pub const MAX: Self = Self(18);
 
     /// Creates a new `ProficiencyBonus` with the given value.
     ///
-    /// In debug mode, this will panic if the value is outside the valid range of 1 to 30.
+    /// In debug mode, this will panic if the value is outside the valid range of 2 to 18.
     ///
     /// In release mode, it will clamp the value to the range of [`Self::MIN`] to [`Self::MAX`].
     #[must_use]
     pub const fn new(value: u8) -> Self {
         debug_assert!(
             !(value < Self::MIN.value() || value > Self::MAX.value()),
-            "Proficiency bonus must be between 2 and 9"
+            "Proficiency bonus must be between 2 and 18"
         );
         Self::new_clamped(value)
     }
@@ -55,12 +56,12 @@ impl ProficiencyBonus {
     ///
     /// # Errors
     ///
-    /// Returns an error if the value is outside the valid range of 2 to 9.
+    /// Returns an error if the value is outside the valid range of 2 to 18.
     pub fn try_new(value: u8) -> Result<Self, &'static str> {
         if value < Self::MIN.value() {
             Err("Proficiency bonus cannot be less than 2")
         } else if value > Self::MAX.value() {
-            Err("Proficiency bonus cannot be greater than 9")
+            Err("Proficiency bonus cannot be greater than 18")
         } else {
             Ok(Self(value))
         }
@@ -105,7 +106,7 @@ mod tests {
 
     #[test]
     fn new_clamped_max() {
-        let bonus = ProficiencyBonus::new_clamped(10);
+        let bonus = ProficiencyBonus::new_clamped(18);
         assert_eq!(bonus, ProficiencyBonus::MAX);
     }
 
@@ -123,8 +124,8 @@ mod tests {
 
     #[test]
     fn try_new_max() {
-        let bonus = ProficiencyBonus::try_new(10);
-        assert_eq!(bonus, Err("Proficiency bonus cannot be greater than 9"));
+        let bonus = ProficiencyBonus::try_new(19);
+        assert_eq!(bonus, Err("Proficiency bonus cannot be greater than 18"));
     }
 
     #[test]
@@ -134,7 +135,7 @@ mod tests {
     }
 
     #[test]
-    #[should_panic(expected = "Proficiency bonus must be between 2 and 9")]
+    #[should_panic(expected = "Proficiency bonus must be between 2 and 18")]
     fn new_panic() {
         let _bonus = ProficiencyBonus::new(1);
     }