+++ src/core/skill.rs
@@ -1,5 +1,6 @@
 use crate::core::Ability;
 use core::{fmt::Display, str::FromStr};
+use enumflags2::bitflags;
 
 /// A category of things creatures try to do with an ability check.
 ///
@@ -12,28 +13,68 @@ use core::{fmt::Display, str::FromStr};
 /// ```rust
 /// use dnd::core::{Skill, Ability};
 ///
-/// let acrobatics = Skill::ACROBATICS;
+/// let acrobatics = Skill::Acrobatics;
 /// assert_eq!(acrobatics.name(), "Acrobatics");
 /// assert_eq!(acrobatics.ability(), Ability::Dexterity);
 /// ```
+#[bitflags]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize), serde(transparent))]
-pub struct Skill {
-    name: &'static str,
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[repr(u32)]
+pub enum Skill {
+    /// Stay on your feet in a tricky situation, or perform an acrobatic stunt.
+    Acrobatics,
 
-    #[cfg_attr(feature = "serde", serde(skip))]
-    ability: Ability,
-}
+    /// Calm or train an animal, or get an animal to behave in a certain way.
+    AnimalHandling,
 
-#[cfg(feature = "serde")]
-impl<'a> serde::Deserialize<'a> for Skill {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'a>,
-    {
-        let name: &'a str = serde::Deserialize::deserialize(deserializer)?;
-        Skill::from_str(name).map_err(serde::de::Error::custom)
-    }
+    /// Recall lore about spells, magic items, and the planes of existence.
+    Arcana,
+
+    /// Jump farther than normal, stay afloat in rough water, or break something.
+    Athletics,
+
+    /// Tell a convincing lie, or wear a disguise convincingly.
+    Deception,
+
+    /// Recall lore about historical events, people, nations, and cultures.
+    History,
+
+    /// Discern a person’s mood and intentions.
+    Insight,
+
+    /// Awe or threaten someone into doing what you want.
+    Intimidation,
+
+    /// Find obscure information in books, or deduce how something works.
+    Investigation,
+
+    /// Diagnose an illness, or determine what killed the recently slain.
+    Medicine,
+
+    /// Recall lore about terrain, plants, animals, and weather.
+    Nature,
+
+    /// Using a combination of senses, notice something that’s easy to miss.
+    Perception,
+
+    /// Act, tell a story, perform music, or dance.
+    Performance,
+
+    /// Honestly and graciously convince someone of something.
+    Persuasion,
+
+    /// Recall lore about gods, religious rituals, and holy symbols.
+    Religion,
+
+    /// Pick a pocket, conceal a handheld object, or perform legerdemain.
+    SleightOfHand,
+
+    /// Escape notice by moving quietly and hiding behind things.
+    Stealth,
+
+    /// Follow tracks, forage, find a trail, or avoid natural hazards.
+    Survival,
 }
 
 impl Skill {
@@ -41,96 +82,72 @@ impl Skill {
     #[must_use]
     pub const fn all() -> &'static [Skill] {
         &[
-            Self::ACROBATICS,
-            Self::ANIMAL_HANDLING,
-            Self::ARCANA,
-            Self::ATHLETICS,
-            Self::DECEPTION,
-            Self::HISTORY,
-            Self::INSIGHT,
-            Self::INTIMIDATION,
-            Self::INVESTIGATION,
-            Self::MEDICINE,
-            Self::NATURE,
-            Self::PERCEPTION,
-            Self::PERFORMANCE,
-            Self::PERSUASION,
-            Self::RELIGION,
-            Self::SLEIGHT_OF_HAND,
-            Self::STEALTH,
-            Self::SURVIVAL,
+            Skill::Acrobatics,
+            Skill::AnimalHandling,
+            Skill::Arcana,
+            Skill::Athletics,
+            Skill::Deception,
+            Skill::History,
+            Skill::Insight,
+            Skill::Intimidation,
+            Skill::Investigation,
+            Skill::Medicine,
+            Skill::Nature,
+            Skill::Perception,
+            Skill::Performance,
+            Skill::Persuasion,
+            Skill::Religion,
+            Skill::SleightOfHand,
+            Skill::Stealth,
+            Skill::Survival,
         ]
-    }
-
-    /// Stay on your feet in a tricky situation, or perform an acrobatic stunt.
-    pub const ACROBATICS: Skill = Skill::new("Acrobatics", Ability::Dexterity);
-
-    /// Calm or train an animal, or get an animal to behave in a certain way.
-    pub const ANIMAL_HANDLING: Skill = Skill::new("Animal Handling", Ability::Wisdom);
-
-    /// Recall lore about spells, magic items, and the planes of existence.
-    pub const ARCANA: Skill = Skill::new("Arcana", Ability::Intelligence);
-
-    /// Jump farther than normal, stay afloat in rough water, or break something.
-    pub const ATHLETICS: Skill = Skill::new("Athletics", Ability::Strength);
-
-    /// Tell a convincing lie, or wear a disguise convincingly.
-    pub const DECEPTION: Skill = Skill::new("Deception", Ability::Charisma);
-
-    /// Recall lore about historical events, people, nations, and cultures.
-    pub const HISTORY: Skill = Skill::new("History", Ability::Intelligence);
-
-    /// Discern a person’s mood and intentions.
-    pub const INSIGHT: Skill = Skill::new("Insight", Ability::Wisdom);
-
-    /// Awe or threaten someone into doing what you want.
-    pub const INTIMIDATION: Skill = Skill::new("Intimidation", Ability::Charisma);
-
-    /// Find obscure information in books, or deduce how something works.
-    pub const INVESTIGATION: Skill = Skill::new("Investigation", Ability::Intelligence);
-
-    /// Diagnose an illness, or determine what killed the recently slain.
-    pub const MEDICINE: Skill = Skill::new("Medicine", Ability::Wisdom);
-
-    /// Recall lore about terrain, plants, animals, and weather.
-    pub const NATURE: Skill = Skill::new("Nature", Ability::Intelligence);
-
-    /// Using a combination of senses, notice something that’s easy to miss.
-    pub const PERCEPTION: Skill = Skill::new("Perception", Ability::Wisdom);
-
-    /// Act, tell a story, perform music, or dance.
-    pub const PERFORMANCE: Skill = Skill::new("Performance", Ability::Charisma);
-
-    /// Honestly and graciously convince someone of something.
-    pub const PERSUASION: Skill = Skill::new("Persuasion", Ability::Charisma);
-
-    /// Recall lore about gods, religious rituals, and holy symbols.
-    pub const RELIGION: Skill = Skill::new("Religion", Ability::Intelligence);
-
-    /// Pick a pocket, conceal a handheld object, or perform legerdemain.
-    pub const SLEIGHT_OF_HAND: Skill = Skill::new("Sleight of Hand", Ability::Dexterity);
-
-    /// Escape notice by moving quietly and hiding behind things.
-    pub const STEALTH: Skill = Skill::new("Stealth", Ability::Dexterity);
-
-    /// Follow tracks, forage, find a trail, or avoid natural hazards.
-    pub const SURVIVAL: Skill = Skill::new("Survival", Ability::Wisdom);
-
-    #[must_use]
-    const fn new(name: &'static str, ability: Ability) -> Self {
-        Self { name, ability }
     }
 
     /// Returns the name of the skill.
     #[must_use]
     pub const fn name(&self) -> &'static str {
-        self.name
+        match self {
+            Skill::Acrobatics => "Acrobatics",
+            Skill::AnimalHandling => "Animal Handling",
+            Skill::Arcana => "Arcana",
+            Skill::Athletics => "Athletics",
+            Skill::Deception => "Deception",
+            Skill::History => "History",
+            Skill::Insight => "Insight",
+            Skill::Intimidation => "Intimidation",
+            Skill::Investigation => "Investigation",
+            Skill::Medicine => "Medicine",
+            Skill::Nature => "Nature",
+            Skill::Perception => "Perception",
+            Skill::Performance => "Performance",
+            Skill::Persuasion => "Persuasion",
+            Skill::Religion => "Religion",
+            Skill::SleightOfHand => "Sleight of Hand",
+            Skill::Stealth => "Stealth",
+            Skill::Survival => "Survival",
+        }
     }
 
     /// Returns the ability associated with this skill.
     #[must_use]
     pub const fn ability(&self) -> Ability {
-        self.ability
+        match self {
+            Skill::Acrobatics | Skill::SleightOfHand | Skill::Stealth => Ability::Dexterity,
+            Skill::Athletics => Ability::Strength,
+            Skill::AnimalHandling
+            | Skill::Insight
+            | Skill::Medicine
+            | Skill::Perception
+            | Skill::Survival => Ability::Wisdom,
+            Skill::Arcana
+            | Skill::History
+            | Skill::Investigation
+            | Skill::Nature
+            | Skill::Religion => Ability::Intelligence,
+            Skill::Deception | Skill::Intimidation | Skill::Performance | Skill::Persuasion => {
+                Ability::Charisma
+            }
+        }
     }
 }
 
@@ -150,11 +167,27 @@ impl FromStr for Skill {
     type Err = &'static str;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Self::all()
-            .iter()
-            .find(|&skill| skill.name() == s)
-            .copied()
-            .ok_or("Unknown skill")
+        match s {
+            "Acrobatics" => Ok(Skill::Acrobatics),
+            "Animal Handling" => Ok(Skill::AnimalHandling),
+            "Arcana" => Ok(Skill::Arcana),
+            "Athletics" => Ok(Skill::Athletics),
+            "Deception" => Ok(Skill::Deception),
+            "History" => Ok(Skill::History),
+            "Insight" => Ok(Skill::Insight),
+            "Intimidation" => Ok(Skill::Intimidation),
+            "Investigation" => Ok(Skill::Investigation),
+            "Medicine" => Ok(Skill::Medicine),
+            "Nature" => Ok(Skill::Nature),
+            "Perception" => Ok(Skill::Perception),
+            "Performance" => Ok(Skill::Performance),
+            "Persuasion" => Ok(Skill::Persuasion),
+            "Religion" => Ok(Skill::Religion),
+            "Sleight of Hand" => Ok(Skill::SleightOfHand),
+            "Stealth" => Ok(Skill::Stealth),
+            "Survival" => Ok(Skill::Survival),
+            _ => Err("Unknown skill"),
+        }
     }
 }
 
@@ -199,7 +232,7 @@ mod tests {
     fn serde() {
         use serde_json;
 
-        let skill = Skill::ACROBATICS;
+        let skill = Skill::Acrobatics;
         let serialized = serde_json::to_string(&skill).unwrap();
         assert_eq!(serialized, "\"Acrobatics\"");