+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-06-22
+
+### Added
+
+- `SkillProficiencies` and `SkillLevel` to track proficiency levels for skills.
+
+### Changed
+
+- `Skill` is now an `enum` instead of a struct with constants.
+- Some types have a specific size representation.
+- Proficiency bonuses can now go up to `18` (to represent expertise)
+- 
+
 ## [0.1.0] - 2025-06-22
 
 ### Added